requirements.txt
dnspython==1.16.0
PyNaCl
async-timeout==3.0.1
discord.py
discord.py[voice]
imdbpy
psycopg2
table2ascii
python-dotenv
deep_translator
cinemagoer
DiscordUtils
gtts
asyncio
ffmpeg
